requirements.txt
click~=8.1.7
psycopg2~=2.9.9
requests~=2.32.3
folium~=0.14.0
grpcio~=1.68.0
protobuf~=5.28.3
tqdm~=4.67.1
pip~=24.2
zlib~=1.2.13
wheel~=0.44.0
pysocks~=1.7.1
openssl~=3.0.15
jinja2~=3.1.4
setuptools~=75.1.0
absl-py~=2.1.0